staff details added

+++ src/main/java/com/smartsecurity/system/dto/VehicleRequest.java
@@ -18,6 +18,6 @@ public class VehicleRequest {
     private String company;
     private String purpose;
     private Long tenantId;
-    private Long createdByUserId;
+    private Integer createdByUserId;
     private UserType userType;
 }
